Add constraint handling and use it for pdf upload

+++ src/EssayTask/PdfInput/Service.php
@@ -4,40 +4,55 @@ declare(strict_types=1);
 
 namespace Edutiek\AssessmentService\EssayTask\PdfInput;
 
-use Edutiek\AssessmentService\EssayTask\Essay\FullService as EssayService;
-use Edutiek\AssessmentService\System\BackgroundTask\Manager as BackgroundTaskManager;
-use Edutiek\AssessmentService\EssayTask\Data\Essay;
-use Edutiek\AssessmentService\EssayTask\EssayImage\Service as EssayImage;
-use Edutiek\AssessmentService\Task\CorrectorComment\FullService as CorrectorComment;
-use Edutiek\AssessmentService\System\Language\FullService as Language;
+use Edutiek\AssessmentService\EssayTask\Api\ApiException;
 use Edutiek\AssessmentService\EssayTask\BackgroundTask\GenerateEssayImages;
-use DateTimeImmutable;
-use Closure;
-use Edutiek\AssessmentService\System\File\Storage;
+use Edutiek\AssessmentService\EssayTask\Data\Essay;
+use Edutiek\AssessmentService\EssayTask\Essay\FullService as EssayService;
+use Edutiek\AssessmentService\EssayTask\EssayImage\Service as EssayImage;
+use Edutiek\AssessmentService\System\BackgroundTask\Manager as BackgroundTaskManager;
+use Edutiek\AssessmentService\System\ConstraintHandling\Actions\ChangeWritingContent;
+use Edutiek\AssessmentService\System\ConstraintHandling\Collector;
+use Edutiek\AssessmentService\System\ConstraintHandling\Result;
+use Edutiek\AssessmentService\System\ConstraintHandling\ResultStatus;
 use Edutiek\AssessmentService\System\EventHandling\Dispatcher;
 use Edutiek\AssessmentService\System\EventHandling\Events\WritingContentChanged;
+use Edutiek\AssessmentService\System\File\Storage;
+use Edutiek\AssessmentService\System\Language\FullService as Language;
 
 readonly class Service implements FullService
 {
-    /**
-     * @param Closure(int): CorrectorComment $get_corrector_comment
-     */
     public function __construct(
+        private bool $as_admin,
         private EssayService $essay_service,
         private BackgroundTaskManager $task_manager,
         private EssayImage $essay_image,
         private Language $language,
         private Storage $storage,
-        private Dispatcher $dispatcher
+        private Dispatcher $events,
+        private Collector $constraints
     ) {
+    }
+
+    public function checkReplacePdf(Essay $essay): Result
+    {
+        return $this->constraints->check(new ChangeWritingContent(
+            $essay->getWriterId(),
+            $essay->getTaskId(),
+            $this->as_admin
+        ));
     }
 
     public function replacePdf(Essay $essay, string $file_id): void
     {
+        $result = $this->checkDeletePdf($essay);
+        if ($result->status() == ResultStatus::BLOCK) {
+            throw new ApiException(implode("/n", $result->messages()), ApiException::CONSTRAINT);
+        }
+
         $this->storage->deleteFile($essay->getPdfVersion());
         $this->essay_image->deleteByEssayId($essay->getId());
         $this->essay_service->save($essay->setPdfVersion($file_id)->touch());
-        $this->dispatcher->dispatchEvent(new WritingContentChanged(
+        $this->events->dispatchEvent(new WritingContentChanged(
             $essay->getWriterId(),
             $essay->getTaskId(),
             $essay->getLastChange()
@@ -51,11 +66,25 @@ readonly class Service implements FullService
         );
     }
 
-    public function deletePdf($essay): void
+    public function checkDeletePdf(Essay $essay): Result
     {
+        return $this->constraints->check(new ChangeWritingContent(
+            $essay->getWriterId(),
+            $essay->getTaskId(),
+            $this->as_admin
+        ));
+    }
+
+    public function deletePdf(Essay $essay): void
+    {
+        $result = $this->checkDeletePdf($essay);
+        if ($result->status() == ResultStatus::BLOCK) {
+            throw new ApiException(implode("/n", $result->messages()), ApiException::CONSTRAINT);
+        }
+
         $this->storage->deleteFile($essay->getPdfVersion());
         $this->essay_service->save($essay->setPdfVersion(null)->touch());
-        $this->dispatcher->dispatchEvent(new WritingContentChanged(
+        $this->events->dispatchEvent(new WritingContentChanged(
             $essay->getWriterId(),
             $essay->getTaskId(),
             $essay->getLastChange()
